Small scoreedit and dssi controls out fix. More separation of Alsa/Jack midi processing -  as a possible fix for crash report by Geoff B. Please see ChangeLog

+++ muse/controlfifo.h
@@ -33,9 +33,9 @@ struct ControlEvent
   //  (possibly rounded) frame and index as the previous item. This is mainly for 
   //  dssi-vst guis, they require acknowledgment of every message.
   bool unique; 
-  unsigned idx;
+  unsigned long idx;
   float value;
-  unsigned frame;    
+  unsigned long frame;    
 };
 
 //---------------------------------------------------------
